# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_vmarble_warehouse-management-service_internal_domain.PlanStatus:
    enum:
    - DRAFT
    - APPROVED
    - CANCELED
    type: string
    x-enum-varnames:
    - PlanDraft
    - PlanApproved
    - PlanCanceled
...
  github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_planning_Plan:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_module_planning.Plan'
        type: array
      limit:
        type: integer
      total_is_estimate:
        type: boolean
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_planning_PlanLookupItem:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_module_planning.PlanLookupItem'
        type: array
      limit:
        type: integer
      total_is_estimate:
        type: boolean
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_module_planning.BoostPriorityResult:
    properties:
      audit_id:
        type: string
      boosted_at:
        type: string
    type: object
  internal_module_planning.CreatePlanInput:
    properties:
      deadline:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_module_planning.PlanItemInput'
        type: array
      po_id:
        type: string
      sales_order_id:
        type: string
    type: object
  internal_module_planning.FeasibilityResult:
    properties:
      feasible:
        type: boolean
      reason:
        type: string
      suggestions:
        items:
          $ref: '#/definitions/internal_module_planning.FeasibilitySuggestion'
        type: array
    type: object
  internal_module_planning.FeasibilitySuggestion:
    properties:
      days_to_due:
        type: integer
      freed_qty:
        type: integer
      score:
        type: number
      sku_code:
        type: string
      wo_id:
        type: string
    type: object
  internal_module_planning.Plan:
    properties:
      canceled_at:
        type: string
      canceled_by:
        type: string
      canceled_reason:
        type: string
      code:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_module_planning.PlanItem'
        type: array
      po_code:
        type: string
      po_id:
        type: string
      sales_order_code:
        type: string
      sales_order_id:
        type: string
      status:
        $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_domain.PlanStatus'
    type: object
  internal_module_planning.PlanItem:
    properties:
      id:
        type: string
      plan_id:
        type: string
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_module_planning.PlanItemInput:
    properties:
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_module_planning.PlanLookupItem:
    properties:
      code:
        type: string
      deadline:
        type: string
      id:
        type: string
      po_code:
        type: string
      sales_order_code:
        type: string
      status:
        $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_domain.PlanStatus'
    type: object
  internal_module_planning.PreemptCandidate:
    properties:
      current_so_code:
        type: string
      freed_qty:
        type: integer
      slack_days:
        type: integer
      status:
        type: string
      wo_id:
        type: string
    type: object
  internal_module_planning.PreemptResult:
    properties:
      audit_id:
        type: string
      freed_qty:
        type: integer
      preempted_at:
        type: string
    type: object
  internal_module_planning.boostPriorityRequest:
    properties:
      reason:
        type: string
    type: object
  internal_module_planning.cancelPlanRequest:
    properties:
      reason:
        type: string
    type: object
  internal_module_planning.preemptRequest:
    properties:
      from_wo_id:
        type: string
      reason:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for warehouse & production management.
  title: VMARBLE Warehouse Management Service API
  version: 0.1.0
paths:
...
  /api/v1/planning/work-orders/{id}/boost-priority:
    post:
      consumes:
      - application/json
      parameters:
      - description: work order id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_module_planning.boostPriorityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_planning.BoostPriorityResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Boost work order priority
      tags:
      - planning
  /api/v1/planning/work-orders/{id}/check-feasibility:
    post:
      parameters:
      - description: work order id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_planning.FeasibilityResult'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check work order material feasibility
      tags:
      - planning
  /api/v1/planning/work-orders/{id}/preempt:
    post:
      consumes:
      - application/json
      parameters:
      - description: target work order id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: from_wo_id + reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_module_planning.preemptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_planning.PreemptResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "412":
          description: Precondition Failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Preempt a work order to free materials for another
      tags:
      - planning
  /api/v1/planning/work-orders/{id}/preempt-candidates:
    get:
      parameters:
      - description: work order id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_module_planning.PreemptCandidate'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List preemption candidates for a work order
      tags:
      - planning
  /api/v1/plans:
    get:
      parameters:
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: search by plan code or PO code (ILIKE)
        in: query
        name: search
        type: string
      - description: 'filter by status: DRAFT, APPROVED, CANCELED'
        in: query
        name: status
        type: string
      - description: inclusive lower bound on created_at (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: inclusive upper bound on created_at (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: 'sort column: created_at, deadline (default created_at)'
        in: query
        name: sort_by
        type: string
      - description: 'sort direction: asc, desc (default desc)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_planning_Plan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List production plans
      tags:
      - planning
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_module_planning.CreatePlanInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_module_planning.Plan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create production plan
      tags:
      - planning
  /api/v1/plans/{id}:
    get:
      parameters:
      - description: plan id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_planning.Plan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get production plan
      tags:
      - planning
  /api/v1/plans/{id}/approve:
    post:
      parameters:
      - description: plan id (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve production plan
      tags:
      - planning
  /api/v1/plans/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancels a DRAFT or APPROVED plan. APPROVED cancels require a non-empty
        reason and cascade-cancel any PLANNED work orders. Refused if any work order
        has progressed past PLANNED.
      parameters:
      - description: plan id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: cancel reason (required when plan is APPROVED)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_module_planning.cancelPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel production plan
      tags:
      - planning
  /api/v1/plans/lookup:
    get:
      description: Lightweight endpoint for async searchable dropdown. Returns plan
        id, code, po_code, status, deadline only (no items).
      parameters:
      - description: ILIKE search on plan code or PO code
        in: query
        name: search
        type: string
      - description: 'filter by status: DRAFT, APPROVED, CANCELED'
        in: query
        name: status
        type: string
      - description: inclusive lower bound on deadline (YYYY-MM-DD)
        in: query
        name: deadline_from
        type: string
      - description: inclusive upper bound on deadline (YYYY-MM-DD)
        in: query
        name: deadline_to
        type: string
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: items per page (default 20, max 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vmarble_warehouse-management-service_internal_platform_httpkit.PagedResult-internal_module_planning_PlanLookupItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lookup plans for async combobox
      tags:
      - planning
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"